stucture: Move Attachmentslist &co to a more logical place
