+++ attachments/utils.py
@@ -167,109 +167,6 @@ class AttachmentSlot(renderable):
         )
 
 
-class AttachmentItem(DocItem):
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        # self.attachment = self.slot.attachment.get_correct_submodel()
-
-    @property
-    def attachment(
-        self,
-    ):
-        return self.slot.attachment.get_correct_submodel()
-
-    def get_link_url(
-        self,
-    ):
-        return self.slot.attachment.get_download_url(
-            self.slot.get_proposal(),
-        )
-
-    def get_filename(
-        self,
-    ):
-        return self.attachment.upload.original_filename
-
-
-class DocList(list):
-    """
-    A list-like of attachments that can output its contents as containers
-    separated per attached objects.
-    """
-
-    def as_containers(self):
-        containers = []
-        per_item = self.per_item()
-        for item in per_item.keys():
-            if type(item) is Proposal:
-                container = self.make_proposal_container(item)
-            else:
-                # This must be a Study object
-                container = Container(_("Traject {}").format(item.order))
-                container.order = 200 + item.order
-            container.items += [self.make_docitem(slot) for slot in per_item[item]]
-            containers.append(container)
-        return sorted(containers, key=lambda c: c.order)
-
-    def per_item(self):
-        items = set([slot.attached_object for slot in self])
-        item_dict = {item: [] for item in items}
-        for slot in self:
-            item_dict[slot.attached_object].append(slot)
-        return item_dict
-
-    def make_proposal_container(self, proposal):
-        container = Container(_("Aanvraag"))
-        container.order = 100
-        # The proposal PDF isn't an attachment, so we add it manually
-        proposal_pdf = DocItem(_("Aanvraag in PDF-vorm"))
-        proposal_pdf.link_url = reverse("proposals:pdf", args=(proposal.pk,))
-        container.items.append(proposal_pdf)
-        return container
-
-    def make_docitem(self, slot):
-        docitem = AttachmentItem(
-            slot.kind.name,
-            slot=slot,
-        )
-        return docitem
-
-
-class AttachmentsList(renderable):
-
-    template_name = "attachments/attachments_list.html"
-
-    def __init__(
-        self,
-        review,
-        request=None,
-    ):
-        self.review = review
-        self.proposal = review.proposal
-        self.containers = self.get_containers()
-        self.request = request
-
-    def get_containers(
-        self,
-    ):
-        from proposals.utils.stepper import Stepper
-
-        stepper = Stepper(self.proposal)
-        filled_slots = [slot for slot in stepper.attachment_slots if slot.attachment]
-        containers = DocList(
-            filled_slots,
-        ).as_containers()
-        return containers
-
-    def get_context_data(self, *args, **kwargs):
-        context = super().get_context_data(*args, **kwargs)
-        context["review"] = self.review
-        context["proposal"] = self.proposal
-        context["containers"] = self.containers
-        return context
-
-
 def get_kind_from_str(db_name):
     from attachments.kinds import ATTACHMENTS
