+++ reviews/utils/__init__.py
@@ -1,2 +1,3 @@
 from .review_actions import *
 from .review_utils import *
+from .attachment_utils import AttachmentsList
